# Main Functionality

This functions are mainly used for opening memory and file databases.

getdatabase

Autonorm.getdatabase — Function

getdatabase()

Arguments

No arguments

Description

Connects to a on-memory-database file and returns a SQLite.DB object.

Output

Examples

julia> db = getdatabase()
SQLite.DB(":memory:")

References

getdatabase(s)

Arguments

s::AbstractString: Path to database file.

Description

Connects to a database file and returns a SQLite.DB object.

Output

Examples

julia> db = getdatabase("test.db")

References